requirements.txt
langchain==1.0.2
langchain-openai==1.0.1
langchain-mcp-adapters>=0.1.0
fastmcp==2.12.5

# A_stock
tushare
efinance

# Live trading
apscheduler>=3.10.0
requests>=2.28.0
python-dotenv>=1.0.0